+++ up42/http/oauth.py
@@ -64,7 +64,7 @@ class Up42Auth(requests.auth.AuthBase):
         self.token_settings = token_settings
         self.adapter = create_adapter(include_post=True)
         self.retrieve = retrieve
-        self._token = self._fetch_token()
+        self.token = self._fetch_token()
         self._lock = threading.Lock()
 
     def __call__(self, request: requests.PreparedRequest) -> requests.PreparedRequest:
@@ -79,9 +79,9 @@ class Up42Auth(requests.auth.AuthBase):
     @property
     def _access_token(self) -> Token:
         with self._lock:
-            if self._token.has_expired:
-                self._token = self._fetch_token()
-        return self._token.access_token
+            if self.token.has_expired:
+                self.token = self._fetch_token()
+        return self.token.access_token
 
     def __deepcopy__(self, memo: dict):
         # Pystac client deep copies the request modifier this class is used for.