+++ CHANGELOG.md
@@ -29,10 +29,14 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 2.1.1a4
+
+**Nov TBD, 2024**
+- Remove the usage of `users/me` endpoint from `base::authenticate`.
+
 ## 2.1.1a3
 
 **Oct 29, 2024**
-### Added
 - Fix `Catalog::construct_search_parameters` `limit` description in the documentation.
 
 ## 2.1.1a2